meta-fso: task-fso2-compliance: drop fsoaudiod as dependency
fsoaudiod is not supported very well as default audio daemon on most machines so it will
be not part of SHR's 2012.01 release which is currently the only potential user of
fsoaudiod. With the release after the 2012.01 one fsoaudiod will become the default audio
daemon in SHR and it should be part of the task-fso2-compliance daemon again.

Signed-off-by: Simon Busch <morphis@gravedo.de>