meta-openmoko: remove gta01 support from 2.6.39 and add 3.6.37 for gta01
Add 2.6.37 kernel for GTA01, remove incorrect build instructions for existing 2.6.39 kernel

The 2.6.39 kernel currently used in shr-core fails to build for GTA01
and, in fact has no code to support the machine.  This commit provides
a 2.6.37 kernel with GTA01 support in order to allow SHR images to be
built.  It also removes om-gta01 as a build target for the 2.6.39
kernel.

This is an intermediate fix; a proper fix requires forward-porting
GTA01 support from this 2.6.37 kernel, ideally to mainline.  The
kernel patch comes from openmoko git, via shr-unstable.

Signed-off-by: Bob Ham <rah@bash.sh>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>