Add a recipe for geeksphone one kernel.
The patch adds a new recipe for the geeksphone one kernel. Kernel
sources are hosted on github and are based on the work of rmcc (Ricardo
Maria Cerqueira). Current kernel version is 2.6.35.7.
At the moment I'm using a ramdisk with a static compiled busybox to boot
SHR from the SD card, I'm not sure if the kernel can boot without the
ramdisk. This ramdisk needs to be created manually, and flashed on the
recovery partition of the phone to boot. I Will try to solve this later
in another patch.

Signed-off-by: Víctor Enríquez <victor.quicksilver@gmail.com>