meta-samsung: libgles-android: rework installation process to copy relevant binaries on first boot
Due to legal reasons we're not able to ship the properitary android binaries anymore. So
we have to copy them from already installed android to the relevant places on first boot.
This requires a specific android version to be installed on the device. This is at least
android 4.1.1 right now the and version of the recipe should track the currently supported
android version.

Signed-off-by: Simon Busch <morphis@gravedo.de>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>