meta-android: android-system-image: correct typo
The typo caused the script not to set the permission bits for any file correctly so the
android init daemon wasn't able to load some files containing required system settings.
In the last consequence this leads to segfaults when loading android libraries through
libhybris as the system wasn't able to retrieve any of the required properties from the
property_service.

Signed-off-by: Simon Busch <morphis@gravedo.de>