meta-lg: base-files: prevent systemd from hardly requiring partitions to be mounted
By default systemd puts in a a RequiredBy dependency on local-fs.target for partitions in
fstab unless you tell it to not do so. On our recent builds both needed partitions for
firmware and persistant data were not available early enough so it steopped the boot
process. With specifying "nofail" as option we tell systemd to continue booting and don't
wait for those partitions to be mounted.

Furthermore we now create the /persist directory which wasn't created before.

Signed-off-by: Simon Busch <morphis@gravedo.de>