cornucopia: inherit from autotools-brokensep to fix Vala builds
Automake's support for Vala does not work with a separate build
directory when using $(top_srcdir) in a _VALAFLAGS variable, like
fsogsmd does.  For some reason, Automake will change directory to
$(srcdir) before running valac:

  http://git.savannah.gnu.org/cgit/automake.git/tree/bin/automake.in#n5474

The $(top_srcdir) variable is relative to the build directory so the
change of directory makes $(top_srcdir) point somewhere wrong.
Automake includes generated C files in the distribution, allowing for
compilation without valac and I suspect this is the root of the
issue.  There is an Automake bug, #13002, filed about this:

  http://debbugs.gnu.org/cgi/bugreport.cgi?bug=13002

To fix our build, we just prevent the use of a separate build
directory by inheriting from autotools-brokensep.

Signed-off-by: Bob Ham <bob.ham@collabora.com>