android-kernel-bootimg: dedicated recipe for creating boot.img
Previous the kernel_android.bbclass inherit was used to
create and deploy the Android boot.img image. However, as the
latter depends on initramfs, it can create dependency loops
if the initramfs includes recipes that rely on the kernel,
like udev of lvm.

This new recipe needs to be included in the final rootfs image.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>