systemd: for arm32 Halium targets, disable network protection for hostnamed
On hammerhead, using Halium-9.0, systemd-hostnamed leads to a crash in
the kernel related to network namespaces.
Backporting various commits to this kernel didn't work, so an easy
workaround is to just disable network isolation for this specific service.

The long term solution will be to switch to a mainline kernel for hammerhead.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>