Revert "android-system-image: use latest image built on jenkins"
This reverts commit 0cd461ace2141f1e9942bf800d96ef7882588dbe.

The images built on jenkins have some issues, revert this until it's resolved
in halium build job.

Booting an image with this update was failing with:
16:41 <+JaMa> [  103.458988] init: Couldn't open /sys/fs/selinux/null: No such file or directory
16:41 <+JaMa> [  103.459062] init: Reboot start, reason: reboot, rebootTarget: bootloader
16:41 <+JaMa> [  103.459189] init: Reboot ending, jumping to kernel

there are couple unexpected differences in the android image:
17:21 <+JaMa> android-system-image-hammerhead# diff -rq *rfs/ 2>/dev/null | grep Only
17:21 <+JaMa> Only in 20210405-1-r0-old-rfs/: init.extraenv.armeabi-v7a.rc
17:21 <+JaMa> Only in 20210429-18-r0-new-rfs/: res
17:21 <+JaMa> Only in 20210405-1-r0-old-rfs/system/etc: recovery-resource.dat
17:21 <+JaMa> Only in 20210405-1-r0-old-rfs/system/lib: lib
17:21 <+JaMa> Only in 20210405-1-r0-old-rfs/system/lib64: lib

we think it's because the halium patches from
https://github.com/Halium/hybris-patches/tree/halium-9.0/hybris/hybris-boot
didn't apply causing
https://github.com/Halium/hybris-patches/blob/halium-9.0/apply-patches.sh
to fail early (due to set -e) and then bunch of other more important
patches weren't applied.

The current plan is to remove
https://github.com/Halium/hybris-patches/tree/halium-9.0/hybris/hybris-boot
either in upstream:
https://github.com/Halium/hybris-patches/pull/20
if all halium users don't use hybris-boot or to use our fork of hybris-patches with:
https://github.com/webOS-ports/hybris-patches/pull/1
and update the manifest in:
https://github.com/webOS-ports/android/blob/luneos-halium-9.0/halium/default.xml#L25

Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>