Update license file and remove patch
The latest version of yocto does not seem to like reaching up one
directory above the package root for the license file, so this uses
the COPYING file instead which has the same contents.

The patch file is also no longer needed because the upstream has
addressed the original problem.

Signed-off-by: Ed Beroset <beroset@ieee.org>