linux-xiaomi-rosy: add unused patchset for sensors
When working on the DTS for the sensors, it can be
useful to list the exact model that is actually
present in the device.

To do that, a set of alpha-quality patches can be
applied on the 6.9.1 msm8953 kernel in order to be
able to query the iio sensors from userland using
the QRTR bus.

The patchset is a rebase from the commits of
https://gitlab.com/msm8996-mainline/linux/-/commits/msm8996-staging-smgr
(c50ea36bf7bcc68b48c0b4051c586fe2f0a7ec35 to 92672b3e6591ea506b49a454d2fcd01b0def34b5)

A recipe 'sns-reg' can be used to do the actual query.
The result can then be seen in the dmesg logs, and
will look like this:

[ 155.041110] smgr 5-8: Getting available sensors
[ 155.043392] smgr 5-8: 0x00,0: BOSCH BMI120 Accelerometer
[ 155.043418] smgr 5-8: 0x00,1: BOSCH BMI120 Temperature
[ 155.044872] smgr 5-8: 0x0a,0: BOSCH BMI120 Gyroscope
[ 155.044894] smgr 5-8: 0x0a,1: BOSCH BMI120 Temperature
[ 155.047186] smgr 5-8: 0x14,0: AKM AK09918 Magnetometer
[ 155.049106] smgr 5-8: 0x28,0: Sensortek stk3x1x alsprx
[ 155.049136] smgr 5-8: 0x28,1: Sensortek stk3x1x alsprx

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>