lk2nd-msm8953: remove useless workaround
lk2nd is actually looking for an "androidboot.hardware=qcom" string in
cmdline to know if it is a downstream kernel or not.
Then, for downstream kernels, it appends the bootloader cmdline to the
kernel cmdline; this also means the DTS will not be patched before boot.

As linux-xiaomi-{rosy,tissot} recipes have been fixed, remove this now
useless workaround.

Signed-off-by: Christophe Chapuis <chris.chapuis@gmail.com>